create zip file for lambda
